+++ src/test/java/analytics/util/ScoringSingletonTest.java
@@ -1363,7 +1363,10 @@ public class ScoringSingletonTest {
 	}
 	@AfterClass
 	public static void cleanUp(){
-		conn.dropDatabase();
+		if(conn.toString().equalsIgnoreCase("FongoDB.test"))
+			conn.dropDatabase();
+			  else
+			   Assert.fail("Something went wrong. Tests connected to " + conn.toString());
 	}
 	
 }
